requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
bcrypt==3.2.0
black==22.3.0
blinker==1.4
boto3==1.18.62
botocore==1.21.62
cffi==1.15.0
click==8.0.3
debugpy==1.5.0
decorator==5.1.0
dnspython==1.16.0
email-validator==1.1.3
entrypoints==0.3
Flask==2.0.2
Flask-Admin==1.5.8
Flask-Classful==0.14.2
flask-mongoengine==1.0.0
flask-swagger==0.2.14
flask-swagger-ui==3.36.0
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==3.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
jupyter-client==7.0.6
jupyter-core==4.8.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mongoengine==0.23.1
mypy-extensions==0.4.3
nest-asyncio==1.5.1
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.1
PyYAML==6.0
pyzmq==22.3.0
regex==2020.1.8
repackage==0.7.3
s3transfer==0.5.0
six==1.16.0
tomli==1.2.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
WTForms==2.3.3
